Curl_checkheaders: make it available for IMAP and SMTP too
... not only HTTP uses this now.
